seperate the code to store in both MyComposable.kt and MyFunction.kt
